feat: Use int64_t for chunk shapes and validate dimensions
This commit refactors the file format and I/O layers to use `int64_t` instead of `uint32_t` for chunk shape dimensions. This change future-proofs the format, allowing for dimension sizes that exceed the `uint32_t` limit.

Key changes:
- The `Chunk` struct in `cdd_file_format.h` now uses `memory::vector<int64_t>` for `shape_`.
- Serialization and deserialization logic in `cdd_file_format.cpp` and `data_writer.cpp` has been updated to handle the 8-byte integer type.
- All I/O and compression APIs (`DataWriter`, `DataReader`, `DataCompressor`, `DataExtractor`) have been updated to accept and process `int64_t` shapes.
- Added robust validation to reject negative shape dimensions at the API boundary in `DataWriter` and `DataCompressor`.
- Added safety checks in `DataReader` and `DataExtractor` to handle potentially corrupt files with negative dimensions gracefully.
- Updated the test suite in `cdd_file_tests.cpp` to use `int64_t` for shape definitions and comparisons.
- Added a new test case, `WriterRejectsNegativeShape`, to explicitly verify that the writer rejects invalid negative dimensions.

+++ src/data_io/data_compressor.h
@@ -59,7 +59,7 @@ class DataCompressor
      */
     [[nodiscard]] ChunkResult compress_zstd(
         std::span<const std::byte> data,
-        std::span<const uint32_t> shape,
+        std::span<const int64_t> shape,
         DType dtype,
         int level = ZstdCompressor::DEFAULT_COMPRESSION_LEVEL
     ) const;
@@ -124,7 +124,7 @@ class DataCompressor
     [[nodiscard]] ChunkResult compress_chunk(
         std::span<const float> data,
         ChunkDataType type,
-        std::span<const uint32_t> shape,
+        std::span<const int64_t> shape,
         std::span<const float> prev_state,
         int level = ZstdCompressor::DEFAULT_COMPRESSION_LEVEL
     ) const;
@@ -141,7 +141,7 @@ class DataCompressor
     [[nodiscard]] ChunkResult compress_chunk(
         std::span<const int64_t> data,
         ChunkDataType type,
-        std::span<const uint32_t> shape,
+        std::span<const int64_t> shape,
         std::span<const int64_t> prev_row,
         int level = ZstdCompressor::DEFAULT_COMPRESSION_LEVEL
     ) const;
